feat: first implementation of creating an invoice
Implementation of the creation of minimal valid XRechnung files in the UBL
format. Calculation of invoice sums and taxes from given hourly work and
configuration of byers and seller via TOML file. This includes
documentation for the crate and a simple command line application that can
create a bill from a CSV input file of recorded hours.